chore: Update mocha
Update mocha from 3.5 to 4.1. This resolves some issues reported by `npm
audit`.

To work with mocha 6, the express test had to explicitly the server.
Mocha no longer forces an exit when it thinks the test is done, so the
server was keeping the process running, causing the test suite to never
finish.